+++ pyonep/exceptions.py
@@ -29,4 +29,6 @@ class ProvisionException(OneException):
     def __init__(self, provision_response):
         self.response = provision_response
     def __str__(self):
+        return self.__repr__()
+    def __repr__(self):
         return "{0} {1}".format(self.response.status(), self.response.reason())