+++ pyonep/onep.py
@@ -35,29 +35,37 @@ except ImportError:
 
 
 class DeferredRequests():
-    '''Encapsulates a list of deferred requests for each CIK. Once the requests
+    '''Encapsulates a list of deferred requests for each auth/CIK. Once the requests
         are ready to be sent, get_method_args_pairs() returns a list of the
         method name and arguments for each request.'''
     def __init__(self):
         self._requests = {}
 
-    def add(self, cik, method, args):
-        '''Append a deferred request for a particular CIK.'''
-        self._requests.setdefault(cik, []).append((method, args))
+    def _authstr(self, auth):
+        '''Convert auth to str so that it can be hashed'''
+        if type(auth) is dict:
+            return '{' + ','.join(["{0}:{1}".format(k, auth[k]) for k in sorted(auth.keys())]) + '}'
+        else:
+            return auth
 
-    def reset(self, cik):
-        self._requests.pop(cik)
+    def add(self, auth, method, args):
+        '''Append a deferred request for a particular auth/CIK.'''
+        self._requests.setdefault(self._authstr(auth), []).append((method, args))
 
-    def has_requests(self, cik):
-        '''Returns True if there are any deferred requests for CIK, False
-        otherwise.'''
-        return (cik in self._requests
-                and len(self._requests[cik]) > 0)
+    def reset(self, auth):
+        self._requests.pop(self._authstr(auth))
 
-    def get_method_args_pairs(self, cik):
+    def has_requests(self, auth):
+        '''Returns True if there are any deferred requests for
+        auth/CIK, False otherwise.'''
+        authstr = self._authstr(auth)
+        return (authstr in self._requests
+                and len(self._requests[authstr]) > 0)
+
+    def get_method_args_pairs(self, auth):
         '''Returns a list of method/arguments pairs corresponding to deferred
-        requests for this CIK'''
-        return self._requests[cik]
+        requests for this auth/CIK'''
+        return self._requests[self._authstr(auth)]
 
 
 class OnepV1():
@@ -100,7 +108,7 @@ class OnepV1():
         '''Returns a list of request bodies made by this instance of OnepV1'''
         return self._loggedrequests
 
-    def _callJsonRPC(self, cik, callrequests, returnreq=False):
+    def _callJsonRPC(self, auth, callrequests, returnreq=False):
         '''Calls the Exosite One Platform RPC API.
             If returnreq is False, result is a tuple with this structure:
                 (success (boolean), response)
@@ -109,7 +117,8 @@ class OnepV1():
             this structure:
                 (request, success, response)
                 '''
-        auth = self._getAuth(cik)
+        # get full auth (auth could be a CIK str)
+        auth = self._getAuth(auth)
         jsonreq = {"auth": auth, "calls": callrequests}
         if self.logrequests:
             self._loggedrequests.append(jsonreq)
@@ -168,13 +177,17 @@ class OnepV1():
 
         raise OneException("Unknown error")
 
-    def _getAuth(self, cik):
+    def _getAuth(self, auth):
         '''Create the authorization/identification portion of a request.'''
-        if None != self._clientid:
-            return {"cik": cik, "client_id": self._clientid}
-        elif None != self._resourceid:
-            return {"cik": cik, "resource_id": self._resourceid}
-        return {"cik": cik}
+        if type(auth) is dict:
+            return auth
+        else:
+            # auth is string
+            if None != self._clientid:
+                return {"cik": auth, "client_id": self._clientid}
+            elif None != self._resourceid:
+                return {"cik": auth, "resource_id": self._resourceid}
+            return {"cik": auth}
 
     def _composeCalls(self, method_args_pairs):
         calls = []
@@ -186,27 +199,27 @@ class OnepV1():
             i += 1
         return calls
 
-    def _call(self, method, cik, arg, defer):
+    def _call(self, method, auth, arg, defer):
         if defer:
-            self.deferred.add(cik, method, arg)
+            self.deferred.add(auth, method, arg)
             return True
         else:
             calls = self._composeCalls([(method, arg)])
-            return self._callJsonRPC(cik, calls)
+            return self._callJsonRPC(auth, calls)
 
-    def has_deferred(self, cik):
-        return self.deferred.has_requests(cik)
+    def has_deferred(self, auth):
+        return self.deferred.has_requests(auth)
 
-    def send_deferred(self, cik):
-        '''Send all deferred requests for a particular cik.'''
-        if self.deferred.has_requests(cik):
+    def send_deferred(self, auth):
+        '''Send all deferred requests for a particular CIK/auth.'''
+        if self.deferred.has_requests(auth):
             calls = self._composeCalls(
-                self.deferred.get_method_args_pairs(cik))
+                self.deferred.get_method_args_pairs(auth))
             try:
-                r = self._callJsonRPC(cik, calls, returnreq=True)
+                r = self._callJsonRPC(auth, calls, returnreq=True)
             finally:
                 # remove deferred calls
-                self.deferred.reset(cik)
+                self.deferred.reset(auth)
             return r
         raise JsonRPCRequestException('No deferred requests to send.')
 
@@ -219,70 +232,70 @@ class OnepV1():
         self._clientid = None
 
     # API methods
-    def activate(self, cik, codetype, code, defer=False):
-        return self._call('activate', cik, [codetype, code], defer)
+    def activate(self, auth, codetype, code, defer=False):
+        return self._call('activate', auth, [codetype, code], defer)
 
-    def create(self, cik, type, desc, defer=False):
-        return self._call('create', cik, [type, desc], defer)
+    def create(self, auth, type, desc, defer=False):
+        return self._call('create', auth, [type, desc], defer)
 
-    def deactivate(self, cik, codetype, code, defer=False):
-        return self._call('deactivate', cik, [codetype, code], defer)
+    def deactivate(self, auth, codetype, code, defer=False):
+        return self._call('deactivate', auth, [codetype, code], defer)
 
-    def drop(self, cik, rid, defer=False):
-        return self._call('drop', cik, [rid], defer)
+    def drop(self, auth, rid, defer=False):
+        return self._call('drop', auth, [rid], defer)
 
-    def flush(self, cik, rid, options=None, defer=False):
+    def flush(self, auth, rid, options=None, defer=False):
         args = [rid]
         if options is not None:
             args.append(options)
-        return self._call('flush', cik, args, defer)
+        return self._call('flush', auth, args, defer)
 
-    def info(self, cik, rid, options={}, defer=False):
-        return self._call('info', cik,  [rid, options], defer)
+    def info(self, auth, rid, options={}, defer=False):
+        return self._call('info', auth,  [rid, options], defer)
 
-    def listing(self, cik, types, options=None, defer=False):
+    def listing(self, auth, types, options=None, defer=False):
         if options is None:
             # This variant is deprecated
-            return self._call('listing', cik, [types], defer)
+            return self._call('listing', auth, [types], defer)
         else:
-            return self._call('listing', cik, [types, options], defer)
+            return self._call('listing', auth, [types, options], defer)
 
-    def lookup(self, cik, type, mapping, defer=False):
-        return self._call('lookup', cik, [type, mapping], defer)
+    def lookup(self, auth, type, mapping, defer=False):
+        return self._call('lookup', auth, [type, mapping], defer)
 
-    def map(self, cik, rid, alias, defer=False):
-        return self._call('map', cik, ['alias', rid, alias], defer)
+    def map(self, auth, rid, alias, defer=False):
+        return self._call('map', auth, ['alias', rid, alias], defer)
 
-    def read(self, cik, rid, options, defer=False):
-        return self._call('read', cik, [rid, options], defer)
+    def read(self, auth, rid, options, defer=False):
+        return self._call('read', auth, [rid, options], defer)
 
-    def record(self, cik, rid, entries, options={}, defer=False):
-        return self._call('record', cik, [rid, entries, options], defer)
+    def record(self, auth, rid, entries, options={}, defer=False):
+        return self._call('record', auth, [rid, entries, options], defer)
 
-    def recordbatch(self, cik, rid, entries, defer=False):
-        return self._call('recordbatch', cik, [rid, entries], defer)
+    def recordbatch(self, auth, rid, entries, defer=False):
+        return self._call('recordbatch', auth, [rid, entries], defer)
 
-    def revoke(self, cik, codetype, code, defer=False):
-        return self._call('revoke', cik, [codetype, code], defer)
+    def revoke(self, auth, codetype, code, defer=False):
+        return self._call('revoke', auth, [codetype, code], defer)
 
-    def share(self, cik, rid, options={}, defer=False):
-        return self._call('share', cik, [rid, options], defer)
+    def share(self, auth, rid, options={}, defer=False):
+        return self._call('share', auth, [rid, options], defer)
 
-    def tag(self, cik, rid, action, tag, defer=False):
-        return self._call('tag', cik, [rid, action, tag], defer)
+    def tag(self, auth, rid, action, tag, defer=False):
+        return self._call('tag', auth, [rid, action, tag], defer)
 
-    def unmap(self, cik, alias, defer=False):
-        return self._call('unmap', cik, ['alias', alias], defer)
+    def unmap(self, auth, alias, defer=False):
+        return self._call('unmap', auth, ['alias', alias], defer)
 
-    def update(self, cik, rid, desc={}, defer=False):
-        return self._call('update', cik, [rid, desc], defer)
+    def update(self, auth, rid, desc={}, defer=False):
+        return self._call('update', auth, [rid, desc], defer)
 
-    def usage(self, cik, rid, metric, starttime, endtime, defer=False):
-        return self._call('usage', cik,
+    def usage(self, auth, rid, metric, starttime, endtime, defer=False):
+        return self._call('usage', auth,
                           [rid, metric, starttime, endtime], defer)
 
-    def write(self, cik, rid, value, options={}, defer=False):
-        return self._call('write', cik, [rid, value, options], defer)
+    def write(self, auth, rid, value, options={}, defer=False):
+        return self._call('write', auth, [rid, value, options], defer)
 
-    def writegroup(self, cik, entries, defer=False):
-        return self._call('writegroup', cik, [entries], defer)
+    def writegroup(self, auth, entries, defer=False):
+        return self._call('writegroup', auth, [entries], defer)